Added check for grasp and double trajectory planning
Planning with both rotations of the arm sometimes saves some seconds.
Added a check for how much the cartesian controller follows the
calculated path. This is usefull when a collision does not allow the arm
to arrive at the tomato. This is a problem with the octomap not
correctly updating probably, but is a good check nonetheless.